Move all task handlers to one module
