+++ aleph/logic/mapping.py
@@ -58,7 +58,7 @@ def map_to_aggregator(collection, mapping, aggregator):
     log.info("[%s] Mapping done (%s rows)", mapping.id, idx)
 
 
-def load_mapping(stage, collection, mapping_id, sync=False):
+def load_mapping(collection, mapping_id, sync=False):
     """Flush and reload all entities generated by a mapping."""
     mapping = Mapping.by_id(mapping_id)
     if mapping is None:
@@ -88,11 +88,7 @@ def load_mapping(stage, collection, mapping_id, sync=False):
         aggregator.close()
 
 
-def op_load_mapping_handler(collection, task):
-    load_mapping(task.stage, collection, **task.payload)
-
-
-def flush_mapping(stage, collection, mapping_id, sync=True):
+def flush_mapping(collection, mapping_id, sync=True):
     """Delete entities loaded by a mapping"""
     log.debug("Flushing entities for mapping: %s", mapping_id)
     origin = mapping_origin(mapping_id)
@@ -101,8 +97,3 @@ def flush_mapping(stage, collection, mapping_id, sync=True):
     aggregator.close()
     delete_entities(collection.id, origin=origin, sync=sync)
     update_collection(collection, sync=sync)
-
-
-def op_flush_mapping_handler(collection, task):
-    sync = task.context.get("sync", False)
-    flush_mapping(task.stage, collection, sync=sync, **task.payload)
