+++ setup.py
@@ -31,14 +31,14 @@ setup(
             "adfs = aleph.oauth:handle_adfs_oauth",
         ],
         "aleph.task_handlers": [
-            "index = aleph.logic.processing:op_index_handler",
-            "xref = aleph.logic.xref:op_xref_handler",
-            "xitem = aleph.logic.xref:op_xref_item_handler",
-            "reingest = aleph.logic.collections:op_reingest_handler",
-            "reindex = aleph.logic.collections:op_reindex_handler",
-            "loadmapping = aleph.logic.mapping:op_load_mapping_handler",
-            "flushmapping = aleph.logic.mapping:op_flush_mapping_handler",
-            "exportsearch = aleph.logic.export:op_export_search_results_handler",
+            "index = aleph.task_handlers:op_index_handler",
+            "xref = aleph.task_handlers:op_xref_handler",
+            "xitem = aleph.task_handlers:op_xref_item_handler",
+            "reingest = aleph.task_handlers:op_reingest_handler",
+            "reindex = aleph.task_handlers:op_reindex_handler",
+            "loadmapping = aleph.task_handlers:op_load_mapping_handler",
+            "flushmapping = aleph.task_handlers:op_flush_mapping_handler",
+            "exportsearch = aleph.task_handlers:op_export_search_results_handler",
         ],
         "console_scripts": ["aleph = aleph.manage:cli"],
     },
